fes61: Remediate PR 935 CodeQL findings
